requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
av==15.1.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comm==0.2.3
comtypes==1.4.11
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
dlib==20.0.0
dnspython==2.7.0
docopt==0.6.2
email-validator==2.3.0
emoji==2.15.0
exceptiongroup==1.3.1
executing==2.2.1
facenet-pytorch==2.6.0
facial-emotion-recognition==0.3.4
fastapi==0.128.6
filelock==3.18.0
flask==3.1.1
flatbuffers==25.12.19
fonttools==4.59.0
fsspec==2025.5.1
gast==0.7.0
gdown==5.2.1
git-filter-repo==2.47.0
gitdb==4.0.12
gitpython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.11
importlib-metadata==8.7.1
importlib-resources==6.5.2
imutils==0.5.4
ipykernel==6.30.1
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.3
Js2Py==0.74
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
keras==3.10.0
kernels==0.9.0
kiwisolver==1.4.7
libclang==18.1.1
lightgbm==4.6.0
markdown==3.9
markdown-it-py==3.0.0
markupsafe==3.0.3
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
ml-dtypes==0.5.4
MouseInfo==0.1.3
mpmath==1.3.0
namex==0.1.0
narwhals==1.48.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
oauthlib==3.3.1
openai==1.101.0
opencv-contrib-python==4.11.0.86
opencv-python==4.13.0.90
opt-einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==2.3.3
parso==0.8.5
pillow==11.3.0
pipwin==0.5.2
platformdirs==4.4.0
progressbar==2.5
prompt-toolkit==3.0.52
protobuf==6.33.5
psutil==7.0.0
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
PyAudio==0.2.14
pyautogui==0.9.54
pydantic==2.12.5
pydantic-core==2.41.5
pydeck==0.9.1
pygame==2.6.1
PyGetWindow==0.0.9
pygments==2.19.2
pygoogle-image==1.0.0
pyjokes==0.8.3
pyjsparser==2.7.1
pymsgbox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyPrind==2.11.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5-sip==12.17.0
PyRect==0.2.0
pyscreeze==1.0.1
pySmartDL==1.3.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-magic-bin==0.4.14
python-multipart==0.0.20
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sklearn==0.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
speechrecognition==3.14.3
stack-data==0.6.3
starlette==0.49.3
streamlit==1.47.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.20.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.2.1
torch==2.2.2
torchvision==0.17.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.39.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
werkzeug==3.1.5
wikipedia==1.4.0
wrapt==2.1.1
xgboost==2.1.4
zipp==3.23.0
